+++ ogmaneo/encoder.py
@@ -56,7 +56,7 @@ class Encoder:
                 num_weights = num_hidden_cells * area * vld.size[2] * vld.size[3]
 
                 vl.weights = cl.clrandom.rand(cq, (num_weights,), np.float32, a=0.99, b=1.0)
-                vl.usages = cl.array.zeros(cq, (num_weights,), np.uint8)
+                vl.usages = cl.array.zeros(cq, (num_weights,), np.int32)
                 vl.reconstruction = cl.array.empty(cq, (num_visible_cells,), np.float32)
 
                 self.vls.append(vl)
@@ -99,7 +99,7 @@ class Encoder:
                 num_weights = num_hidden_cells * area * vld.size[2] * vld.size[3]
 
                 vl.weights = cl.array.empty(cq, (num_weights,), np.float32)
-                vl.usages = cl.array.empty(cq, (num_weights,), np.uint8)
+                vl.usages = cl.array.empty(cq, (num_weights,), np.int32)
                 vl.reconstruction = cl.array.empty(cq, (num_visible_cells,), np.float32)
 
                 read_into_buffer(fd, vl.weights)
